+++ docs/src/Fulma/Elements/Icon/View.fs
@@ -8,6 +8,17 @@ open Types
 open Fulma.Elements
 open Fulma.Extra.FontAwesome
 
+let icon =
+    div [ ClassName "block" ]
+        [ Icon.icon [ Icon.isSmall ]
+            [ i [ ClassName "fa fa-home" ] [ ] ]
+          Icon.icon [ ]
+            [ i [ ClassName "fa fa-lg fa-home" ] [ ] ]
+          Icon.icon [ Icon.isMedium ]
+            [ i [ ClassName "fa fa-2x fa-home" ] [ ] ]
+          Icon.icon [ Icon.isLarge ]
+            [ i [ ClassName "fa fa-3x fa-home" ] [ ] ] ]
+
 //Display modification of container size with Bulma options
 let containerSizes =
     div [ ClassName "block" ]
@@ -40,155 +51,132 @@ let containerSizes =
 //Diplay Font Awesome Rotation & Flip
 let iconRotationFlip =
     div [ ClassName "block" ]
-        [
-            ul [] [
-                li [] [
-                    Icon.faIcon [ Icon.isMedium ] [
-                        Fa.icon Fa.I.Shield;
-                        //Large icon
-                        Fa.faLg  ]
-                    str "No Rotation"
-                ]
-                li [] [
-                    Icon.faIcon [ Icon.isMedium ] [
-                        Fa.icon Fa.I.Shield;
-                        //Rotate 90 degrees
-                        Fa.rotate90;
-                        //Large icon
-                        Fa.faLg  ]
-                    str "90 degrees rotation"
-                ]
-                li [] [
-                    Icon.faIcon [ Icon.isMedium ] [
-                        Fa.icon Fa.I.Shield;
-                        //Rotation 180 degrees
-                        Fa.rotate180;
-                        Fa.faLg ]
-                    str "180 degrees rotation"
-                ]
-                li [] [
-
-                    Icon.faIcon [ Icon.isMedium ] [
-                        Fa.icon Fa.I.Shield
-                        //Flip Horizontal
-                        Fa.flipHorizontal
-                        Fa.faLg
-                    ]
-                    str "Horizontal flip"
-                ]
-                li [] [
-                    Icon.faIcon [ Icon.isMedium ] [
-                        Fa.icon Fa.I.Shield
-                        //Flip Vertical
-                        Fa.flipVertical
-                        Fa.faLg
-                    ]
-                    str "Vertical flip"
-                ]
-            ]
-       ]
+        [ ul [ ]
+             [ li [ ]
+                  [ Icon.faIcon [ Icon.isMedium ]
+                                [ Fa.icon Fa.I.Shield
+                                  Fa.faLg  ]
+                    str "No Rotation" ]
+               li [ ]
+                  [ Icon.faIcon [ Icon.isMedium ]
+                                [ Fa.icon Fa.I.Shield
+                                  //Rotate 90 degrees
+                                  Fa.rotate90
+                                  //Large icon
+                                  Fa.faLg  ]
+                    str "90 degrees rotation" ]
+               li [ ]
+                  [ Icon.faIcon [ Icon.isMedium ]
+                                [ Fa.icon Fa.I.Shield
+                                  //Rotation 180 degrees
+                                  Fa.rotate180
+                                  Fa.faLg ]
+                    str "180 degrees rotation" ]
+               li [ ]
+                  [ Icon.faIcon [ Icon.isMedium ]
+                                [ Fa.icon Fa.I.Shield
+                                  //Flip Horizontal
+                                  Fa.flipHorizontal
+                                  Fa.faLg ]
+                    str "Horizontal flip" ]
+               li [ ]
+                  [ Icon.faIcon [ Icon.isMedium ]
+                                [ Fa.icon Fa.I.Shield
+                                  //Flip Vertical
+                                  Fa.flipVertical
+                                  Fa.faLg ]
+                    str "Vertical flip" ] ] ]
 
 //Display Font Awesome Animations
 let iconAnimations =
     div [ ClassName "block" ]
-        [
-            ul [] [
-                li [] [
-                    Icon.faIcon [ Icon.isLarge ] [
-                        //Animations work well on Spinner
-                        Fa.icon Fa.I.Spinner;
-                        //Pulse Animation
-                        Fa.pulse;
-                        //Icon 2x times larger
-                        Fa.fa2x  ]
-                    str "Pulse animation"
-                ]
-                li [] [
-                    Icon.faIcon [ Icon.isLarge ] [
-                        //Animations work well wit Cog
-                        Fa.icon Fa.I.Cog;
-                        //Spin animation
-                        Fa.spin;
-                        //Icon 2x times larger
-                        Fa.fa2x  ]
-                    str "Spin animation"
-                ]
-            ]
-       ]
+        [ ul [ ]
+             [ li [ ]
+                  [ Icon.faIcon [ Icon.isLarge ]
+                                [ //Animations work well on Spinner
+                                  Fa.icon Fa.I.Spinner
+                                  //Pulse Animation
+                                  Fa.pulse
+                                  //Icon 2x times larger
+                                  Fa.fa2x ]
+                    str "Pulse animation" ]
+               li [ ]
+                  [ Icon.faIcon [ Icon.isLarge ]
+                                [ //Animations work well wit Cog
+                                  Fa.icon Fa.I.Cog
+                                  //Spin animation
+                                  Fa.spin
+                                  //Icon 2x times larger
+                                  Fa.fa2x ]
+                    str "Spin animation" ] ] ]
+
 //Stacked Icons
 let stackedIcons =
     div [ ClassName "block" ]
-        [
-            ul [] [
-                li [] [
-                    Icon.stackParent [ Fa.Parent.faLg ] [
-                        Icon.stackChild [ Fa.Child.faStack2x; Fa.Child.icon Fa.I.SquareO ]
-                        Icon.stackChild [ Fa.Child.faStack1x; Fa.Child.icon Fa.I.Twitter ]
-                    ]
-                    str "Twitter logo over a square with round corners"
+        [ ul [ ]
+             [ li [ ]
+                  [ Icon.stackParent [ Fa.Parent.faLg ]
+                                     [ Icon.stackChild [ Fa.Child.faStack2x; Fa.Child.icon Fa.I.SquareO ]
+                                       Icon.stackChild [ Fa.Child.faStack1x; Fa.Child.icon Fa.I.Twitter ] ]
+                    str "Twitter logo over a square with round corners" ]
+               li [ ]
+                  [ Icon.stackParent [ Fa.Parent.faLg ]
+                                     [ Icon.stackChild [ Fa.Child.faStack2x; Fa.Child.icon Fa.I.Circle ]
+                                       Icon.stackChild [ Fa.Child.faStack1x; Fa.Child.icon Fa.I.Flag; Fa.Child.colorInverse ] ]
+                    str "One flag with inversed color over a circle" ] ] ]
 
-                ]
-                li [] [
-                    Icon.stackParent [ Fa.Parent.faLg ] [
-                        Icon.stackChild [ Fa.Child.faStack2x; Fa.Child.icon Fa.I.Circle ]
-                        Icon.stackChild [ Fa.Child.faStack1x; Fa.Child.icon Fa.I.Flag; Fa.Child.colorInverse ]
-                    ]
-                    str "One flag with inversed color over a circle"
-                ]
-            ]
-        ]
 //Display Font Awesome Icon List
 let iconList =
     div [ ClassName "block" ]
-        [
-            // faUl creates an unordered list with icons instead of the classic bullet points
-            Icon.faUl [
-                // faLi creates one list item. It has the same options as faIcon
-                Icon.faLi [ Fa.icon Fa.I.CheckSquare ]       [ str "Item done"]
-                Icon.faLi [ Fa.icon Fa.I.Spinner; Fa.spin ]  [ str "Item processing"]
-            ]
-        ]
+        [ // fa_ul creates an unordered list with icons instead of the classic bullet points
+          Icon.fa_ul [ ]
+            [ li [ ]
+                [ Icon.faIcon [ ]
+                              [ Fa.icon Fa.I.CheckSquare
+                                Fa.isLi ]
+                  str "Item done" ]
+              li [ ]
+                [ Icon.faIcon [ ]
+                              [ Fa.icon Fa.I.Spinner
+                                Fa.spin
+                                Fa.isLi ]
+                  str "Item processing" ] ] ]
 
 let borderPulledIcons =
     div [ ClassName "block" ]
-        [
-            span [] [
-                    Icon.faIcon [] [
-                            Fa.icon Fa.I.QuoteLeft
-                            Fa.pullLeft
-                            Fa.border
-                        ]
-                    str "...Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed non risus.\
+        [ span [ ]
+               [ Icon.faIcon [ ]
+                             [ Fa.icon Fa.I.QuoteLeft
+                               Fa.pullLeft
+                               Fa.border ]
+                 str "...Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed non risus.\
                      Suspendisse lectus tortor, dignissim sit amet, adipiscing nec, ultricies sed, dolor.\
                      Cras elementum ultrices diam. Maecenas ligula massa, varius a, semper congue, euismod non, mi.\
-                     Proin porttitor, orci nec nonummy molestie, enim est eleifend mi, non fermentum diam nisl sit amet erat.."
-            ]
-        ]
+                     Proin porttitor, orci nec nonummy molestie, enim est eleifend mi, non fermentum diam nisl sit amet erat.." ] ]
 
 let fontAwesomeIcons =
     div [ ClassName "block" ]
-        [ Icon.faIcon [ Icon.isSmall ] [ Fa.icon Fa.I.Home; Fa.faLg ]
-          Icon.faIcon [ ] [ Fa.icon Fa.I.Tags]
-          Icon.faIcon [ Icon.isMedium ] [ Fa.icon Fa.I.``500px``; Fa.fa4x ]
-          Icon.faIcon [ Icon.isLarge ] [ Fa.icon Fa.I.Android; Fa.fa3x ]  ]
+        [ Icon.faIcon [ Icon.isSmall ] [ Fa.icon Fa.I.Home ]
+          Icon.faIcon [ ] [ Fa.icon Fa.I.Tags; Fa.faLg ]
+          Icon.faIcon [ Icon.isMedium ] [ Fa.icon Fa.I.``500px``; Fa.fa2x ]
+          Icon.faIcon [ Icon.isLarge ] [ Fa.icon Fa.I.Android; Fa.fa3x ] ]
 
 let composeButtons =
     div [ ClassName "block" ]
-        [
-            Button.button [ Button.isDanger ]
-                    [
-                        Icon.faIcon [ Icon.customClass "icopn-size" ] [ Fa.icon Fa.I.Trash; Fa.faLg ]
-                        span [] [ str "  Delete" ]
-                    ]
-            Button.button [ Button.isInfo ]
-                    [
-                        Icon.faIcon [ ] [ Fa.icon Fa.I.User; Fa.fw ]
-                        span [] [ str "User" ]
-                    ]
-        ]
+        [ Button.button [ Button.isDanger ]
+                        [ Icon.faIcon [ ]
+                                      [ Fa.icon Fa.I.Trash; Fa.faLg ]
+                          span [] [ str "  Delete" ] ]
+          Button.button [ Button.isInfo ]
+                        [ Icon.faIcon [ ]
+                                      [ Fa.icon Fa.I.User; Fa.fw ]
+                          span [] [ str "User" ] ] ]
 
 let root model dispatch =
     Render.docPage [ Render.contentFromMarkdown model.Intro
+                     Render.docSection
+                        "### Sizes"
+                        (Viewer.View.root icon model.IconViewer (IconViewerMsg >> dispatch))
                      Render.docSection
                         """
 ### Convenience functions
@@ -223,7 +211,7 @@ If the icon you want to use isn't accessible via `Fa.I.*` please *[open an issue
 You can also use `Fa.I.Custom "fa-my-icon"` as a fix.
 
 ```fsharp
-    // If the "fa-thumbs-up" icon was missing, you could use Fa.Custom to get it:
+    // If the "fa-thumbs-up" icon was missing, you could use Fa.I.Custom to get it:
     Icon.faIcon [ Icon.isLarge ] [ Fa.I.Custom "fa-thumbs-up" ]
 ```
                         """